Adds custom clipboard value
- This commits adds a new prop that allows to pass a custom value to be
copied to the clipboard... There are some cases when a custom component
is used and we don't want to have all of the childrens copied to the
clipboard.

Example: a custom component that renders the following display...

<b>Phone</b> <span>+5211111111</span>
Phone is being display as a result of a dropdown, but only the number
should be copied to the display.

To achieve that we could pass the clipboardValue prop as:
clipboardValue={(values) => {
  return (
    <div key="custom-component-values">{values.nickname} {values.lastName}</div>
  );
}}

The value of the select is not rendered in there.